fix: resolved errors
