+++ packages/builder/providers/src/shortcuts/index.ts
@@ -1,1 +1,1 @@
-export * from "./provider";
+export * from "./wrappers";
