+++ packages/builder/providers/src/shortcuts/wrappers/CanvasShortcuts.tsx
@@ -1,0 +1,256 @@
+"use client";
+import { Toast } from "@rafty/ui";
+import { useCallback, useEffect, type PropsWithChildren } from "react";
+import toast from "react-hot-toast";
+import { useHotkeys } from "react-hotkeys-hook";
+import { useCanvas } from "../../canvas";
+import { useEventBus } from "../../events";
+import { EditorEventListenerProps } from "../../types";
+import { EditorEvent } from "../../utils";
+import { ActionType, useStack } from "../useStack";
+
+const EDITOR_EVENT_STACK = [
+  EditorEvent.BLOCK_ADDITION,
+  EditorEvent.BLOCK_UPDATION,
+  EditorEvent.BLOCK_ID_UPDATION,
+  EditorEvent.BLOCK_DELETION,
+  EditorEvent.BLOCK_REPOSITION,
+  EditorEvent.BLOCK_DUPLICATION,
+  EditorEvent.SCHEMA_RESET,
+];
+
+export function ShortcutsWrapper({ children }: PropsWithChildren) {
+  // const [, copyToClipboard] = useCopyToClipboard();
+  const { schema, select, updateId, update, add, remove, move, set } =
+    useCanvas(
+      ({ schema, select, updateId, remove, add, update, move, set }) => ({
+        schema,
+        select,
+        update,
+        remove,
+        move,
+        add,
+        updateId,
+        set,
+      }),
+    );
+
+  // Undo & Redo
+  const resolveStackActions = useCallback(
+    ({ action, data }: { action: ActionType; data: unknown }) => {
+      const event_type = (data as { event_type: EditorEvent }).event_type;
+
+      if (event_type === EditorEvent.BLOCK_ID_UPDATION) {
+        const { currentBlockId, newBlockId } =
+          data as EditorEventListenerProps[EditorEvent.BLOCK_ID_UPDATION];
+
+        if (action === ActionType.UNDO)
+          updateId({
+            currentBlockId: newBlockId,
+            newBlockId: currentBlockId,
+            shouldEmit: false,
+          });
+        else
+          updateId({
+            currentBlockId,
+            newBlockId,
+            shouldEmit: false,
+          });
+      }
+
+      if (event_type === EditorEvent.BLOCK_ADDITION) {
+        const { blockData, blockId = "" } =
+          data as EditorEventListenerProps[EditorEvent.BLOCK_ADDITION];
+
+        if (action === ActionType.UNDO)
+          remove({
+            blockIds: blockId,
+            shouldEmit: false,
+          });
+        else add({ blockData, blockId, shouldEmit: false });
+      }
+
+      if (event_type === EditorEvent.BLOCK_UPDATION) {
+        const { id, updatedValues, oldValues } =
+          data as EditorEventListenerProps[EditorEvent.BLOCK_UPDATION];
+
+        if (action === ActionType.UNDO)
+          update({ blockId: id, updatedValues: oldValues, shouldEmit: false });
+        else update({ blockId: id, updatedValues, shouldEmit: false });
+      }
+
+      if (event_type === EditorEvent.BLOCK_DELETION) {
+        const { blockIds, changes } =
+          data as EditorEventListenerProps[EditorEvent.BLOCK_DELETION];
+
+        if (action === ActionType.UNDO)
+          changes.forEach((change, index) =>
+            add({
+              blockData: change.block,
+              blockId: blockIds[index],
+              shouldEmit: false,
+              insertionIndex: change.index,
+            }),
+          );
+        else remove({ blockIds, shouldEmit: false });
+      }
+
+      if (event_type === EditorEvent.BLOCK_REPOSITION) {
+        const { sourceBlockId, targetBlockId } =
+          data as EditorEventListenerProps[EditorEvent.BLOCK_REPOSITION];
+
+        if (action === ActionType.UNDO)
+          move({
+            sourceBlockId: targetBlockId,
+            targetBlockId: sourceBlockId,
+            shouldEmit: false,
+          });
+        else
+          move({
+            sourceBlockId,
+            targetBlockId,
+            shouldEmit: false,
+          });
+      }
+
+      if (event_type === EditorEvent.BLOCK_DUPLICATION) {
+        const { newId, block } =
+          data as EditorEventListenerProps[EditorEvent.BLOCK_DUPLICATION];
+
+        if (action === ActionType.UNDO)
+          remove({
+            blockIds: newId,
+            shouldEmit: false,
+          });
+        else add({ blockId: newId, blockData: block, shouldEmit: false });
+      }
+
+      if (event_type === EditorEvent.SCHEMA_RESET) {
+        const { prev, cur } =
+          data as EditorEventListenerProps[EditorEvent.SCHEMA_RESET];
+
+        if (action === ActionType.UNDO) set({ func: () => prev });
+        else set({ func: () => cur });
+      }
+    },
+    [add, remove, set, update, updateId, move],
+  );
+
+  const stack = useStack<unknown>(resolveStackActions);
+  const addEvent = useEventBus((state) => state.add);
+
+  useEffect(() => {
+    // Actions
+    for (const event of EDITOR_EVENT_STACK) {
+      addEvent(event, stack.pushAction);
+    }
+  }, []);
+
+  useHotkeys(
+    "mod+z",
+    () => {
+      toast.custom((t) => (
+        <Toast title="Undo!" severity="success" visible={t.visible} />
+      ));
+
+      stack.undo();
+    },
+    {
+      description: "Undo",
+    },
+  );
+
+  useHotkeys(
+    "mod+y",
+    () => {
+      toast.custom((t) => (
+        <Toast title="Redo!" severity="success" visible={t.visible} />
+      ));
+
+      stack.redo();
+    },
+    {
+      description: "Redo",
+    },
+  );
+
+  useHotkeys("mod+g", () => console.log("Group"), {
+    description: "Group",
+    preventDefault: true,
+  });
+
+  useHotkeys(
+    "mod+a",
+    () =>
+      select({
+        selectedBlockIds: schema.map(({ id }) => id),
+      }),
+    {
+      description: "Select All",
+      preventDefault: true,
+    },
+  );
+
+  useHotkeys(
+    "mod+d",
+    () =>
+      select({
+        selectedBlockIds: null,
+      }),
+    {
+      description: "Select All",
+      preventDefault: true,
+    },
+  );
+
+  // useHotkeys(
+  //   "mod+c",
+  //   (event) => {
+  //     if (active.canvas && active.block) {
+  //       const block = get(active.canvas, active.block);
+  //       if (block) {
+  //         event.preventDefault();
+  //         copyToClipboard(JSON.stringify(block));
+  //       }
+  //     }
+  //   },
+  //   { description: "Copy" },
+  //   [active],
+  // );
+
+  // useHotkeys(
+  //   "mod+x",
+  //   (event) => {
+  //     if (active.canvas && active.block) {
+  //       const block = get(active.canvas, active.block);
+  //       if (block) {
+  //         event.preventDefault();
+  //         copyToClipboard(JSON.stringify(block));
+  //         remove(active.canvas, active.block);
+  //       }
+  //     }
+  //   },
+  //   { description: "Cut" },
+  //   [active],
+  // );
+
+  // // Paste listener
+  // useEffect(() => {
+  //   const onPaste = (event: ClipboardEvent) => {
+  //     if (event.clipboardData) {
+  //       const data = JSON.parse(
+  //         event.clipboardData.getData("text/plain") ?? "",
+  //       );
+  //       if (data) add(active.canvas ?? "nodes", data);
+  //     }
+  //   };
+
+  //   window.addEventListener("paste", onPaste);
+
+  //   return () => {
+  //     window.removeEventListener("paste", onPaste);
+  //   };
+  // }, [add, active]);
+
+  return children;
+}
