+++ packages/builder/providers/src/shortcuts/wrappers/index.ts
@@ -1,0 +1,2 @@
+export * from "./CanvasShortcuts";
+export * from "./WorkspaceShortcuts";
