+++ packages/builder/shared/package.json
@@ -20,6 +20,7 @@
     "reactflow": "^11.10.4",
     "react-icons": "^5.0.1",
     "react-hotkeys-hook": "^4.5.0",
+    "react-hot-toast": "^2.4.1",
     "prism-react-renderer": "^2.3.1",
     "zustand": "^4.5.0"
   },
